Suppord deleting old ads

+++ Controllers/ProductsController.cs
@@ -162,7 +162,8 @@ namespace sellnet.Controllers
                 if (result.Error != null)
                     BadRequest("Can't Delete Product");
             }
-            await _photoService.DeletePhotoAsync(product.ThumbnailId);
+            if (product.ThumbnailId != null) // Check because old ads don't have a thumbnailurl
+                await _photoService.DeletePhotoAsync(product.ThumbnailId);
             _dbContext.Products.Remove(product);
             if (await _dbContext.SaveChangesAsync() > 0)
                 return Ok(new { Message = "Product Deleted successfully" });
